+++ NEWS.md
@@ -7,6 +7,8 @@
     
 * File name in _Find all Files_ is now styled (#36, #43)
 
+* Added support for terminal color theming via `rsthemes_terminal_colors()` (#63)
+
 # rsthemes 0.2.0 (2020-12-07)
 
 ## New Themes