+++ NEWS.md
@@ -7,6 +7,13 @@
     
 * File name in _Find all Files_ is now styled (#36, #43)
 
+* Added 4 new Material Themes: Dark, Lighter, Ocean, and Palenight 
+  (@lusignan, @grrrck #64, #49, #47)
+  
+* Fixed menu item hover and selected colors in complete menu themes
+
+* Added {rsthemes} version and header to final `.rstheme` files
+
 # rsthemes 0.2.0 (2020-12-07)
 
 ## New Themes