+++ NEWS.md
@@ -15,7 +15,8 @@
 * Added {rsthemes} version and header to final `.rstheme` files
 
 * Added support for terminal color theming via `rsthemes_terminal_colors()` 
-  (@nsgrantham #63)
+  (@nsgrantham #63). Most themes now set the first eight terminal colors
+  (black, red, green, yellow, blue, magenta, cyan, white).
 
 # rsthemes 0.2.0 (2020-12-07)